Move templating utils
